more weapons, no levels tho

+++ app/src/main/java/com/example/projectcyber/GameActivity/gameObjects/Projectile/ProjectileMovement.java
@@ -3,5 +3,5 @@ package com.example.projectcyber.GameActivity.gameObjects.Projectile;
 import com.example.projectcyber.GameActivity.GameView;
 
 public abstract class ProjectileMovement {
-    public abstract void update(double deltaTime, GameView gameView, Projectile projectile);
+    public abstract void update(long deltaTime, GameView gameView, Projectile projectile);
 }
